split workoutstartview into more manageable parts
